+++ tests/testthat/test-keyATMHeterogeneity.R
@@ -7,7 +7,7 @@ skip_on_cran()
 library(quanteda)
 library(magrittr)
 
-if (!require(dplyr)) {
+if (!"dplyr" %in% rownames(installed.packages())) {
   skip("dplyr is not installed")
 }
 
@@ -35,19 +35,19 @@ keywords <- list(Government     = c("laws", "law", "executive"),
 vars <- docvars(data_corpus_inaugural)
 vars %>%
   tibble::as_tibble() %>%
-  mutate(Period = case_when(Year <= 1899 ~ "18_19c",
-                            TRUE ~ "20_21c")) %>%
-  mutate(Party = case_when(Party == "Democratic" ~ "Democratic",
-                           Party == "Republican" ~ "Republican",
-                           TRUE ~ "Other")) %>%
-  select(Party, Period) -> vars_selected
+  dplyr::mutate(Period = dplyr::case_when(Year <= 1899 ~ "18_19c",
+                                          TRUE ~ "20_21c")) %>%
+  dplyr::mutate(Party = dplyr::case_when(Party == "Democratic" ~ "Democratic",
+                                         Party == "Republican" ~ "Republican",
+                                         TRUE ~ "Other")) %>%
+  dplyr::select(Party, Period) -> vars_selected
 
 # Set the base line
 vars_selected %>%
-  mutate(Party  = factor(Party, 
-                         levels = c("Other", "Republican", "Democratic")),
-         Period = factor(Period, 
-                         levels = c("18_19c", "20_21c"))) -> vars_selected
+  dplyr::mutate(Party  = factor(Party, 
+                                levels = c("Other", "Republican", "Democratic")),
+                Period = factor(Period, 
+                                levels = c("18_19c", "20_21c"))) -> vars_selected
 
 out <- keyATM(docs              = keyATM_docs,
               no_keyword_topics = 5,
@@ -78,7 +78,7 @@ test_that("Doc Topic", {
   strata_topic <- by_strata_DocTopic(out, by_var = "Period20_21c",
                                           labels = c("18_19c", "20_21c"),
                                           posterior_mean = TRUE)
-  p <- plot(strata_topic, var_name = "Period", topics = c(1,2,3,4))
+  p <- plot(strata_topic, var_name = "Period", show_topic = c(1,2,3,4))
   expect_s3_class(p, "keyATM_fig")
   skip_on_os("linux")
   expect_equal(summary(strata_topic)[[2]]$Point[1], 0.07443097, tolerance = 0.001)